backend/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.8.0rc2
Django==3.2.4
django-debug-toolbar==4.2.0
django-filter==21.1
django-js-asset==2.2.0
django-mptt==0.15.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.0
drf-yasg==1.21.7
idna==3.6
inflection==0.5.1
oauthlib==3.2.2
packaging==23.2
Pillow==10.1.0
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
psycopg2==2.9.9
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
